requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
PyJWT